Always set model = 'netfront' for xen over libvirt.
This will prevent emulated network driver to be attached to virtual machine.
Emulated network driver is not working for xen over libvirt as tap interface
name is suffixed by xen with '-emu' which makes this interface name too long.

Additionaly security groups will also not work for this interface as openstack
is setting up filters on interface without -emu suffix.

Change-Id: I0907ce148e473edf6b81ea782fac2fa6a1b71a2d
Fixes: bug #1202199